Services update (7280), UX writing (7210) and bug corrections (7300, 7211)

+++ apps/builder/components/shared/Graph/Nodes/ItemNode/ItemNodeContent/contents/WhatsAppOptionsContent/WhatsAppOptionsNodeContent.tsx
@@ -25,14 +25,11 @@ export const WhatsAppOptionsNodeContent = ({
 }: Props) => {
   const { deleteItem, updateItem, createItem, updateStep } = useTypebot()
   const [initialContent] = useState(item.content ?? '')
-  const [itemValue, setItemValue] = useState(
-    item.content ?? 'Clique para editar'
-  )
+  const [itemValue, setItemValue] = useState(item.content ?? 'Editar opção')
   const editableRef = useRef<HTMLDivElement | null>(null)
 
   useEffect(() => {
-    if (itemValue !== item.content)
-      setItemValue(item.content ?? 'Clique para editar')
+    if (itemValue !== item.content) setItemValue(item.content ?? 'Editar opção')
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [item])
 
@@ -44,8 +41,7 @@ export const WhatsAppOptionsNodeContent = ({
   }
 
   const handleKeyPress = (e: React.KeyboardEvent<HTMLDivElement>) => {
-    if (e.key === 'Escape' && itemValue === 'Clique para editar')
-      deleteItem(indices)
+    if (e.key === 'Escape' && itemValue === 'Editar opção') deleteItem(indices)
     if (e.key === 'Enter' && itemValue !== '' && initialContent === '')
       handlePlusClick()
   }
@@ -83,7 +79,7 @@ export const WhatsAppOptionsNodeContent = ({
       >
         <EditablePreview
           w="full"
-          color={item.content !== 'Clique para editar' ? 'inherit' : 'gray.500'}
+          color={item.content !== 'Editar opção' ? 'inherit' : 'gray.500'}
           cursor="pointer"
           px={4}
           py={2}
